+++ packages/liveblocks-react-ui/src/components/AiChat.tsx
@@ -4,7 +4,11 @@ import type {
   CopilotId,
   MessageId,
 } from "@liveblocks/core";
-import { RegisterAiTool, useAiChatMessages } from "@liveblocks/react";
+import {
+  RegisterAiKnowledge,
+  RegisterAiTool,
+  useAiChatMessages,
+} from "@liveblocks/react";
 import { useLatest } from "@liveblocks/react/_private";
 import {
   type ComponentProps,
@@ -109,6 +113,11 @@ export interface AiChatProps extends ComponentProps<"div"> {
    * The layout of the chat and its composer.
    */
   layout?: "inset" | "compact";
+
+  /**
+   * The time, in milliseconds, before an AI response will timeout.
+   */
+  responseTimeout?: number;
 
   /**
    * Override the component's strings.
@@ -216,16 +225,20 @@ const AiChatMessages = forwardRef<HTMLDivElement, AiChatMessagesProps>(
         const trailingSpacer = trailingSpacerRef.current;
         const bottomTrailingMarker = bottomTrailingMarkerRef.current;
 
-        let containerHeight: number | null = null;
-        let footerHeight: number | null = null;
-        let messagesHeight: number | null = null;
+        let containerHeight: number | undefined = undefined;
+        let footerHeight: number | undefined = undefined;
+        let messagesHeight: number | undefined = undefined;
 
         const resetTrailingSpace = () => {
           trailingSpacer?.style.removeProperty("height");
           bottomTrailingMarker?.style.removeProperty("top");
         };
 
-        const resizeObserver = new ResizeObserver((entries) => {
+        const updateTrailingSpace = (
+          updatedContainerHeight?: number,
+          updatedFooterHeight?: number,
+          updatedMessagesHeight?: number
+        ) => {
           if (!trailingSpacer || !bottomTrailingMarker) {
             return;
           }
@@ -233,51 +246,31 @@ const AiChatMessages = forwardRef<HTMLDivElement, AiChatMessagesProps>(
           const lastMessage = messages.lastElementChild;
           const penultimateMessage = lastMessage?.previousElementSibling;
 
-          // If there's no last pair of messages, there's no need for any trailing space.
-          if (!lastMessage || !penultimateMessage) {
-            resetTrailingSpace();
-            return;
+          if (updatedContainerHeight === undefined) {
+            updatedContainerHeight = container.getBoundingClientRect().height;
           }
 
-          // If the container's height is based on its content, the container isn't scrollable and there's no need for any trailing space.
-          if (container.scrollHeight === container.clientHeight) {
-            resetTrailingSpace();
-            return;
+          if (updatedFooterHeight === undefined) {
+            updatedFooterHeight = footer.getBoundingClientRect().height;
           }
 
-          let updatedContainerHeight: number | null = containerHeight;
-          let updatedFooterHeight: number | null = footerHeight;
-          let updatedMessagesHeight: number | null = messagesHeight;
-
-          for (const entry of entries) {
-            const entryHeight =
-              entry.borderBoxSize?.[0]?.blockSize ?? entry.contentRect.height;
-
-            if (entry.target === container) {
-              updatedContainerHeight = entryHeight ?? null;
-            } else if (entry.target === footer) {
-              updatedFooterHeight = entryHeight ?? null;
-            } else if (entry.target === messages) {
-              updatedMessagesHeight = entryHeight ?? null;
-            }
+          if (updatedMessagesHeight === undefined) {
+            updatedMessagesHeight = messages.getBoundingClientRect().height;
           }
 
-          // If we don't have all the heights, we can't compute the trailing space.
-          if (
-            updatedContainerHeight === null ||
-            updatedFooterHeight === null ||
-            updatedMessagesHeight === null
-          ) {
-            resetTrailingSpace();
-            return;
-          }
-
-          // If none of the heights have changed, we don't need to do anything.
+          // If the heights haven't changed, there's no need to update the trailing space.
           if (
             updatedContainerHeight === containerHeight &&
             updatedFooterHeight === footerHeight &&
             updatedMessagesHeight === messagesHeight
           ) {
+            if (
+              !lastMessage ||
+              !penultimateMessage ||
+              container.scrollHeight === container.clientHeight
+            ) {
+              resetTrailingSpace();
+            }
             return;
           }
 
@@ -285,6 +278,18 @@ const AiChatMessages = forwardRef<HTMLDivElement, AiChatMessagesProps>(
           containerHeight = updatedContainerHeight;
           footerHeight = updatedFooterHeight;
           messagesHeight = updatedMessagesHeight;
+
+          // If there's no last pair of messages, there's no need for any trailing space.
+          if (!lastMessage || !penultimateMessage) {
+            resetTrailingSpace();
+            return;
+          }
+
+          // If the container isn't scrollable, there's no need for trailing space.
+          if (container.scrollHeight === container.clientHeight) {
+            resetTrailingSpace();
+            return;
+          }
 
           // A
           const penultimateMessageScrollMarginTop = Number.parseFloat(
@@ -313,11 +318,39 @@ const AiChatMessages = forwardRef<HTMLDivElement, AiChatMessagesProps>(
           // Offset what "the bottom" is to the "scroll at the bottom" detection logic,
           // so that it doesn't include the trailing space.
           bottomTrailingMarker.style.top = `${-trailingSpace}px`;
+        };
+
+        const resizeObserver = new ResizeObserver((entries) => {
+          let updatedContainerHeight: number | undefined = containerHeight;
+          let updatedFooterHeight: number | undefined = footerHeight;
+          let updatedMessagesHeight: number | undefined = messagesHeight;
+
+          for (const entry of entries) {
+            const entryHeight =
+              entry.borderBoxSize?.[0]?.blockSize ?? entry.contentRect.height;
+
+            if (entry.target === container) {
+              updatedContainerHeight = entryHeight;
+            } else if (entry.target === footer) {
+              updatedFooterHeight = entryHeight;
+            } else if (entry.target === messages) {
+              updatedMessagesHeight = entryHeight;
+            }
+          }
+
+          updateTrailingSpace(
+            updatedContainerHeight,
+            updatedFooterHeight,
+            updatedMessagesHeight
+          );
         });
 
         resizeObserver.observe(container);
         resizeObserver.observe(footer);
         resizeObserver.observe(messages);
+
+        // Initialize before the first resize.
+        requestAnimationFrame(() => updateTrailingSpace());
 
         return () => {
           resizeObserver.disconnect();
@@ -425,6 +458,7 @@ export const AiChat = forwardRef<HTMLDivElement, AiChatProps>(
       layout = "inset",
       components,
       className,
+      responseTimeout,
       ...props
     },
     forwardedRef
@@ -462,14 +496,14 @@ export const AiChat = forwardRef<HTMLDivElement, AiChatProps>(
       (behavior: "instant" | "smooth", includeTrailingSpace = false) => {
         if (includeTrailingSpace) {
           // Scroll to the bottom marker to include the trailing space,
-          // and wait for a frame in case the trailing space hasn't
+          // but wait for the next tick in case the trailing space hasn't
           // been updated yet. (e.g. when sending a new message)
-          requestAnimationFrame(() => {
+          setTimeout(() => {
             bottomMarkerRef.current?.scrollIntoView({
               behavior,
               block: "end",
             });
-          });
+          }, 0);
         } else {
           // Scroll to the trailing space marker to only scroll to the
           // bottom of the messages, without including the trailing space.
@@ -494,6 +528,16 @@ export const AiChat = forwardRef<HTMLDivElement, AiChatProps>(
         {Object.entries(tools).map(([name, tool]) => (
           <RegisterAiTool key={name} chatId={chatId} name={name} tool={tool} />
         ))}
+
+        {localKnowledge
+          ? localKnowledge.map((knowledge, index) => (
+              <RegisterAiKnowledge
+                key={`${index}:${knowledge.description}`}
+                chatId={chatId}
+                {...knowledge}
+              />
+            ))
+          : null}
 
         <div className="lb-ai-chat-content">
           {isLoading ? (
@@ -530,6 +574,7 @@ export const AiChat = forwardRef<HTMLDivElement, AiChatProps>(
                 data-trailing-spacer=""
                 style={{
                   pointerEvents: "none",
+                  height: 0,
                 }}
                 aria-hidden
               />
@@ -561,7 +606,7 @@ export const AiChat = forwardRef<HTMLDivElement, AiChatProps>(
             copilotId={copilotId as CopilotId}
             overrides={overrides}
             autoFocus={autoFocus}
-            knowledge={localKnowledge}
+            responseTimeout={responseTimeout}
             onComposerSubmit={onComposerSubmit}
             onComposerSubmitted={({ id }) => setLastSentMessageId(id)}
             className={cn(