+++ packages/liveblocks-react/src/types/ai.ts
@@ -1,6 +1,8 @@
 import type {
+  AiAssistantContentPart,
   AiKnowledgeSource,
   AiOpaqueToolDefinition,
+  Relax,
 } from "@liveblocks/core";
 
 export type RegisterAiKnowledgeProps = AiKnowledgeSource & {
@@ -10,6 +12,12 @@ export type RegisterAiKnowledgeProps = AiKnowledgeSource & {
    * precedence.
    */
   id?: string;
+
+  /**
+   * When provided, the knowledge source will only be available for this chatId.
+   * If not provided, this knowledge source will be available globally.
+   */
+  chatId?: string;
 };
 
 export type RegisterAiToolProps = {
@@ -18,8 +26,7 @@ export type RegisterAiToolProps = {
 
   /**
    * When provided, the tool will only be available for this chatId. If not
-   * provided, this tool will globally be made available to any AiChat
-   * instance.
+   * provided, this tool will be available globally.
    */
   chatId?: string;
 
@@ -31,3 +38,19 @@ export type RegisterAiToolProps = {
    */
   enabled?: boolean;
 };
+
+/**
+ * Simplified status for the requested chat.
+ * This hook offers a convenient way to update the UI while an AI chat
+ * generation is in progress.
+ */
+export type AiChatStatus = Relax<
+  | { status: "loading" }
+  | { status: "idle" }
+  | { status: "generating" } // Still generating, but there is no content yet
+  | {
+      status: "generating";
+      partType: Exclude<AiAssistantContentPart["type"], "tool-invocation">;
+    }
+  | { status: "generating"; partType: "tool-invocation"; toolName: string }
+>;